@@ -7,7 +7,7 @@
 	<name>trip-service-kata</name>
 
    <properties>
-        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.source>11</maven.compiler.source>
         <maven.compiler.target>${maven.compiler.source}</maven.compiler.target>
         <hamcrest.version>1.3</hamcrest.version>
         <junit.version>5.3.1</junit.version>
